+++ sso/settings.go
@@ -59,7 +59,7 @@ type Settings struct {
 	PromptColors      PromptColors          `koanf:"PromptColors" yaml:"PromptColors,omitempty"` // go-prompt colors
 	LogLevel          string                `koanf:"LogLevel" yaml:"LogLevel,omitempty"`
 	LogLines          bool                  `koanf:"LogLines" yaml:"LogLines,omitempty"`
-	HistoryLimit      int                   `koanf:"HistoryLimit" yaml:"HistoryLimit,omitempty"`
+	HistoryLimit      int64                 `koanf:"HistoryLimit" yaml:"HistoryLimit,omitempty"`
 	HistoryMinutes    int64                 `koanf:"HistoryMinutes" yaml:"HistoryMinutes,omitempty"`
 	ListFields        []string              `koanf:"ListFields" yaml:"ListFields,omitempty"`
 }
